update listings page, price list
