+++ main/templates/main/listings.html
@@ -33,8 +33,14 @@
                     {% else %}
                     <p class="ml-5 p-0 m-0 text-center pb-2">
                         Here are all of the items that traders in Torn Exchange are buying.
-                        If you are a seller, filter the search by item name, price or rating, to find the best deal for
-                        you!
+                        If you are a seller, filter the search by item name, price or rating, to find the best deal for you!
+                    </p>
+                    <p class="ml-5 p-0 m-0 text-center pb-2">
+                        <b>Note when filtering by status:</b> Idle traders can be just as online as Online traders, they may have switched tabs or having 
+                        Torn opened but doing something away from keyboard so if you use that filter, make sure to always check Idle traders as well.
+                    </p>
+                    <p class="ml-5 p-0 m-0 text-center pb-2">
+                        When choosing a trader, make sure they traded recently - you can check their activity on their price list page.
                     </p>
                     {% endif %}
                 </div>
@@ -56,12 +62,11 @@
 
                                         <div class="p-2 col-2 text-left">
                                             {% if listing.owner.activity_status == 'Online' %}
-                                            <img src="{% static '/main/images/crown-online.png' %}" style="width:30%;">
+                                                <img src="{% static '/main/images/crown-online.png' %}" style="width:30%;" title="Online" alt="Online">
                                             {% elif listing.owner.activity_status == 'Idle' %}
-                                            <img src="{% static '/main/images/crown-idle.png' %}" style="width: 30%;">
+                                                <img src="{% static '/main/images/crown-idle.png' %}" style="width: 30%;" title="Idle" alt="Idle">
                                             {% elif listing.owner.activity_status == 'Offline' %}
-                                            <img src="{% static '/main/images/crown-offline.png' %}"
-                                                style="width: 30%;">
+                                                <img src="{% static '/main/images/crown-offline.png' %}" style="width: 30%;" title="Offline", alt="Offline">
                                             {% endif %}
                                         </div>
                                         <div class="col text-left">
@@ -117,15 +122,11 @@
                                     <img src="{{listing.item.image_url}}" style="height:50%; width:auto;">
                                 </div>
                             </div>
-
-
-
-
                         </div>
                         <div class="card-footer mt-0">
                             <div class="row">
                                 <div class="col">
-                                    Last updated: {{listing.last_updated|naturaltime}}
+                                    <i class="fas fa-user"></i> Player last active: {{listing.owner.last_active|time_since}}
                                 </div>
                             </div>
                         </div>
